+++ src/pages/v7/settings/language/component.js
@@ -1,11 +1,18 @@
 define(function(require) {
 	
 	var template = require("template7!./template.html");
-
-
+	var $ = require("jquery");
+	
 	return {
+		on: {
+			pageInit: function(e) {
+				e.target.down(
+					"#" + window.locale.loc + " .item-after").innerHTML = "√";
+			}
+		},
+		
 		bindings: {
-			".languages .item-content click": function(e) {
+			".item-content click": function(e) {
 				var id = this.id;
 				$("li .item-after", this.parentNode).html("");
 				$(".item-after", this).html("√");
@@ -16,6 +23,7 @@ define(function(require) {
 				}, 200);
 			}
 		},
+		
 		template: template
 	};