+++ src/pages/veldoffice/onderzoek/new-popup/component.js
@@ -1,0 +1,6 @@
+define(function(require) {
+	
+	var template = require("template7!./template.html");
+	
+	return { template: template };
+});